Move _load_keyspec() to new xmlsec.crypto, and make thread local cache.
